Refactoring, code- and documentation cleanup
